+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonColorTable2013.cs
@@ -541,7 +541,7 @@ public class KryptonColorTable2013 : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => _colors[(int)SchemeBaseColors.StatusStripText];
+    public override Color ToolStripText => _colors.ResolveToolStripText(SchemeBaseColors.StatusStripText);
 
     #endregion