+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/KryptonVisualStudio2010With2010ColorTable.cs
@@ -498,7 +498,7 @@ public class KryptonVisualStudio2010With2010ColorTable : KryptonColorTable
     /// <summary>
     /// Gets the text color used on the tool strip.
     /// </summary>
-    public override Color ToolStripText => Colors[(int)SchemeBaseColors.StatusStripText];
+    public override Color ToolStripText => Colors.ResolveToolStripText(SchemeBaseColors.StatusStripText);
 
     #endregion